requirements.txt
absl-py==2.1.0
ace==0.3.3
ace-tools==0.0
ace-tools-open==0.1.0
altgraph==0.17
asttokens==2.4.1
astunparse==1.6.3
audioset-download==1.0
backcall==0.2.0
black==24.8.0
brotli==1.1.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
contourpy==1.0.5
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
executing==2.1.0
filelock==3.16.1
flatbuffers==24.3.25
fsspec==2024.9.0
future==0.18.2
gast==0.4.0
google-auth==2.34.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.66.1
h5py==3.11.0
idna==3.8
importlib-metadata==8.5.0
importlib-resources==6.4.0
ipykernel==6.29.5
ipython==8.12.2
itables==2.2.3
jedi==0.19.1
jinja2==3.1.4
joblib==1.4.2
keras==2.13.1
kiwisolver==1.4.4
libclang==18.1.1
markdown==3.7
markupsafe==2.1.5
matplotlib==3.7.2
matplotlib-inline==0.1.7
mpmath==1.3.0
mutagen==1.47.0
mypy-extensions==1.0.0
networkx==3.1
numpy==1.24.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==24.1
pandas==2.0.3
parso==0.8.4
pathspec==0.12.1
pefile==2019.4.18
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.3.2
protobuf==4.25.4
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycryptodomex==3.20.0
pygments==2.18.0
pyparsing==3.0.9
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
six==1.16.0
stack-data==0.6.2
sympy==1.13.3
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
threadpoolctl==3.5.0
tomli==2.0.1
torch==2.4.1
torchaudio==2.4.1
tornado==6.4.1
traitlets==5.14.3
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.1
unicodedata2==15.1.0
urllib3==2.2.3
websockets==13.0.1
werkzeug==3.0.4
wrapt==1.16.0
youtube-dl==2021.12.17
yt-dlp==2024.8.6
zipp==3.20.1